@@ -2,8 +2,10 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
+using AutoMapper;
 using OpenData.Domain.Models;
 using OpenData.Domain.Services;
+using OpenData.Resources;
 using OpenData.Services;
 
 namespace OpenData.Controllers
@@ -12,17 +14,20 @@ namespace OpenData.Controllers
 	public class MunicipalityController : Controller
 	{
 		private readonly IMunicipalityService _municipalityService;
+		private readonly IMapper _mapper;
 
-		public MunicipalityController(IMunicipalityService municipalityService) 
+		public MunicipalityController(IMunicipalityService municipalityService, IMapper mapper) 
 		{
 			_municipalityService = municipalityService;
+			_mapper = mapper;
 		}
 
 		[HttpGet]
-		public async Task<IEnumerable<Municipality>> GetAllAsync()
+		public async Task<IEnumerable<MunicipalityResource>> GetAllAsync()
 		{
 			var municipalities = await _municipalityService.ListAsync();
-			return municipalities;
+			var resources = _mapper.Map<IEnumerable<Municipality>, IEnumerable<MunicipalityResource>>(municipalities);
+			return resources;
 		}
 	}
 }
